refactor: 대시보드 관련 컴포넌트 리네이밍
